pyproject.toml
[project]
name = "ah2k25-final-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.15",
    "azure-cognitiveservices-speech>=1.45.0",
    "fastapi>=0.116.1",
    "google-generativeai>=0.8.5",
    "python-dotenv>=1.0.0",
    "joblib>=1.5.1",
    "numpy>=2.3.2",
    "pandas>=2.3.1",
    "pydub>=0.25.1",
    "pymupdf>=1.26.3",
    "pytesseract>=0.3.13",
    "python-multipart>=0.0.20",
    "requests>=2.32.4",
    "scikit-learn>=1.7.1",
    "sentence-transformers>=2.6.0",
    "uvicorn>=0.35.0",
    "xgboost>=3.0.4",
    "dotenv>=0.9.9",
]
